refactor: transport MyDeck to be molecule
